// src/types/collections.d.ts
export interface InvType {
  basePrice: number | 'None'
  capacity: number
  description: string
  graphicID: number
  groupID: number
  iconID: number | 'None'
  marketGroupID: number | 'None'
  mass: number
  portionSize: number
  published: number
  raceID: number | 'None'
  soundID: number | 'None'
  typeID: number
  typeName: string
  volume: number
}
export type InvTypeDocument = WithId<InvType>
// ...
export type OrderStats = {
  weightedAverage: string
  max: string
  min: string
  stddev: string
  median: string
  volume: string
  orderCount: string
  percentile: string
}

export type MarketCache = {
  typeId: number
  structureId: string
  buy: OrderStats
  sell: OrderStats
}

export type MarketCacheDocument = WithID<MarketCache>